requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
python-dotenv
langchain
langchain-community
pypdf
pinecone-client
sentence-transformers
transformers
numpy
tiktoken
google-generativeai
scipy
scikit-learn
tokenizers
huggingface-hub
sqlalchemy
psycopg2-binary
pgvector
